requirements.txt
celery==5.2.7
Django==3.2.16
#django-filter==2.4.0
django-settings-export==1.2.1
djangorestframework==3.14.0
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==5.2.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
PyJWT==2.5.0
pymongo==4.3.2
pymemcache==3.5.2
pytz==2022.5
PyYAML==6.0
gunicorn==20.1.0
